refactor: move gitToplevel beside isGitRepo in utils/git
