+++ packages/cli/src/utils/git.ts
@@ -15,6 +15,20 @@ export function isGitRepo(cwd: string): boolean {
   return exists(nodePath.join(cwd, '.git'));
 }
 
+/** The git worktree root at/above cwd, or undefined outside any repository. */
+export function gitToplevel(cwd: string): string | undefined {
+  try {
+    const top = execFileSync('git', ['rev-parse', '--show-toplevel'], {
+      cwd,
+      encoding: 'utf8',
+      stdio: ['ignore', 'pipe', 'pipe'],
+    }).trim();
+    return top === '' ? undefined : top;
+  } catch {
+    return undefined;
+  }
+}
+
 /**
  * Stop tracking files that are now gitignored but were committed before the
  * ignore rule existed (e.g. safeword's transient hook state). `git rm --cached`
